refactor(mcp): split http server routes helpers
